SO-4731: Move applicable test cases from QL request test to ECL
