+++ snomed/com.b2international.snowowl.snomed.datastore.tests/src/com/b2international/snowowl/snomed/core/ecl/SnomedEclEvaluationRequestTest.java
@@ -24,6 +24,7 @@ import static com.b2international.snowowl.snomed.datastore.index.entry.SnomedCon
 import static com.b2international.snowowl.snomed.datastore.index.entry.SnomedConceptDocument.Expressions.statedAncestors;
 import static com.b2international.snowowl.snomed.datastore.index.entry.SnomedConceptDocument.Expressions.statedParents;
 import static com.b2international.snowowl.test.commons.snomed.DocumentBuilders.*;
+import static com.b2international.snowowl.test.commons.snomed.RandomSnomedIdentiferGenerator.generateDescriptionId;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.fail;
@@ -31,6 +32,8 @@ import static org.junit.Assume.assumeFalse;
 
 import java.math.BigDecimal;
 import java.util.*;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
 
 import org.eclipse.xtext.parser.IParser;
 import org.eclipse.xtext.serializer.ISerializer;
@@ -62,6 +65,7 @@ import com.b2international.snowowl.snomed.core.tree.Trees;
 import com.b2international.snowowl.snomed.datastore.config.SnomedCoreConfiguration;
 import com.b2international.snowowl.snomed.datastore.index.entry.SnomedConceptDocument;
 import com.b2international.snowowl.snomed.datastore.index.entry.SnomedDescriptionIndexEntry;
+import com.b2international.snowowl.snomed.datastore.index.entry.SnomedDocument;
 import com.b2international.snowowl.snomed.datastore.index.entry.SnomedRefSetMemberIndexEntry;
 import com.b2international.snowowl.snomed.datastore.index.entry.SnomedRelationshipIndexEntry;
 import com.b2international.snowowl.snomed.datastore.request.SnomedRequests;
@@ -69,6 +73,7 @@ import com.b2international.snowowl.test.commons.snomed.RandomSnomedIdentiferGene
 import com.b2international.snowowl.test.commons.snomed.TestBranchContext;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSet;
 import com.google.inject.Injector;
 
@@ -167,11 +172,6 @@ public class SnomedEclEvaluationRequestTest extends BaseRevisionIndexTest {
 				.build();
 	}
 	
-	@Test(expected = BadRequestException.class)
-	public void syntaxErrorsShouldThrowException() throws Exception {
-		eval("invalid");
-	}
-	
 	private Expression eval(String expression) {
 		return new RevisionIndexReadRequest<>(SnomedRequests.prepareEclEvaluation(expression)
 				.setExpressionForm(isInferred() ? Trees.INFERRED_FORM : Trees.STATED_FORM) // use the isInferred method decide on inferred vs stated form (this will provide support for axioms as well)
@@ -179,12 +179,16 @@ public class SnomedEclEvaluationRequestTest extends BaseRevisionIndexTest {
 				.execute(context)
 				.getSync();		
 	}
-	
+
+	@Test(expected = BadRequestException.class)
+	public void syntaxError() throws Exception {
+		eval("invalid");
+	}
+
 	@Test(expected = BadRequestException.class)
 	public void _null() throws Exception {
 		eval(null);
 	}
-	
 
 	@Test
 	public void empty() throws Exception {
@@ -424,6 +428,14 @@ public class SnomedEclEvaluationRequestTest extends BaseRevisionIndexTest {
 		final Expression actual = eval(String.format("%s OR %s OR %s", ROOT_ID, ROOT_ID, ROOT_ID));
 		assertNotNull(actual);
 	}
+	
+	@Test
+	public void largeIdSetWithOrOperators() throws Exception {
+		eval(IntStream.range(0, 10_000)
+			.mapToObj(i -> RandomSnomedIdentiferGenerator.generateConceptId())
+			.collect(Collectors.joining(" OR ")));
+	}
+
 	
 	@Test(expected = BadRequestException.class)
 	public void binaryOperatorAmbiguityOrAnd() throws Exception {
@@ -1228,6 +1240,266 @@ public class SnomedEclEvaluationRequestTest extends BaseRevisionIndexTest {
 		assertEquals(expected, actual);
 	}
 	
+	@Test
+	public void filterActiveOnly() throws Exception {
+		final Expression actual = eval("* {{ active=true }}");
+		final Expression expected = SnomedDocument.Expressions.active();
+		assertEquals(expected, actual);
+	}
+	
+	@Test
+	public void filterInactiveOnly() throws Exception {
+		final Expression actual = eval("* {{ active=false }}");
+		final Expression expected = SnomedDocument.Expressions.inactive();
+		assertEquals(expected, actual);
+	}
+	
+	@Test
+	public void filterModuleId() throws Exception {
+		final Expression actual = eval("* {{ moduleId= " + Concepts.MODULE_SCT_CORE + " }}");
+		final Expression expected = SnomedDocument.Expressions.modules(ImmutableList.of(Concepts.MODULE_SCT_CORE));
+		assertEquals(expected, actual);
+	}
+	
+	@Test
+	public void filterTerm() throws Exception {
+		indexRevision(MAIN, SnomedDescriptionIndexEntry.builder()
+			.id(generateDescriptionId())
+			.active(true)
+			.moduleId(Concepts.MODULE_SCT_CORE)
+			.term("Clinical finding")
+			.conceptId(Concepts.ROOT_CONCEPT)
+			.typeId(Concepts.SYNONYM)
+			.build());
+		
+		final Expression actual = eval("* {{ term = \"Clin find\" }}");
+		final Expression expected = SnomedDocument.Expressions.ids(ImmutableList.of(Concepts.ROOT_CONCEPT));
+		assertEquals(expected, actual);
+	}
+	
+	@Test(expected = BadRequestException.class)
+	public void filterTermLessThanTwoChars() throws Exception {
+		eval("* {{ term = \"C\" }}");
+	}
+	
+	@Test
+	public void filterConjunctionActiveAndModuleId() throws Exception {
+		final Expression actual = eval("* {{ active = true, moduleId = " + Concepts.MODULE_SCT_CORE + " }}");
+		final Expression expected = Expressions.builder()
+			.filter(SnomedDocument.Expressions.active())
+			.filter(SnomedDocument.Expressions.modules(ImmutableList.of(Concepts.MODULE_SCT_CORE)))
+			.build();
+		assertEquals(expected, actual);
+	}
+	
+	@Test
+	public void filterDisjunctionActiveAndModuleId() throws Exception {
+		final Expression actual = eval("* {{ active = true OR moduleId = " + Concepts.MODULE_SCT_CORE + " }}");
+		final Expression expected = Expressions.builder()
+			.should(SnomedDocument.Expressions.active())
+			.should(SnomedDocument.Expressions.modules(ImmutableList.of(Concepts.MODULE_SCT_CORE)))
+			.build();
+		assertEquals(expected, actual);
+	}
+	
+	@Test(expected = BadRequestException.class)
+	public void filterConjunctionDomainInconsistency() throws Exception {
+		eval("* {{ active=true AND Description.moduleId = "+ Concepts.MODULE_SCT_CORE +" }}");
+	}
+	
+	@Test(expected = BadRequestException.class)
+	public void filterDisjunctionDomainInconsistency() throws Exception {
+		eval("* {{ Description.active=true OR moduleId = "+ Concepts.MODULE_SCT_CORE +" }}");
+	}
+	
+	@Test
+	public void filterDescriptionType() throws Exception {
+		indexRevision(MAIN, SnomedDescriptionIndexEntry.builder()
+			.id(generateDescriptionId())
+			.active(true)
+			.moduleId(Concepts.MODULE_SCT_CORE)
+			.term("Clinical finding")
+			.conceptId(Concepts.ROOT_CONCEPT)
+			.typeId(Concepts.TEXT_DEFINITION)
+			.build());
+		
+		final Expression actual = eval("* {{ typeId = " + Concepts.TEXT_DEFINITION + " }}");
+		final Expression expected = SnomedDocument.Expressions.ids(ImmutableList.of(Concepts.ROOT_CONCEPT));
+		assertEquals(expected, actual);
+	}
+	
+	@Test(expected = BadRequestException.class)
+	public void filterConjunctionAmbiguity() throws Exception {
+		eval("* {{ Description.active=true AND Description.moduleId = " + Concepts.MODULE_SCT_CORE + " OR term=\"clinical finding\" }}");
+	}
+	
+	@Test(expected = BadRequestException.class)
+	public void filterDisjunctionAmbiguity() throws Exception {
+		eval("* {{ Description.active=true OR Description.moduleId = " + Concepts.MODULE_SCT_CORE + " AND term=\"clinical finding\" }}");
+	}
+	
+	@Test(expected = BadRequestException.class)
+	public void filterExclusionAmbiguity() throws Exception {
+		eval("* {{ Description.active=true OR Description.moduleId = " + Concepts.MODULE_SCT_CORE +" MINUS term=\"clinical finding\" }}");
+	}
+	
+	@Test
+	public void filterMultiDomainQueryAnd() throws Exception {
+		Expression actual = eval("* {{ active=false }} AND * {{ term=\"clin find\" }}");
+		Expression expected = Expressions.builder()
+			.filter(SnomedDocument.Expressions.inactive())
+			.filter(SnomedDocument.Expressions.ids(Collections.emptySet()))
+			.build();
+		assertEquals(expected, actual);
+	}
+	
+	@Test
+	public void filterMultiDomainQueryOr() throws Exception {
+		Expression actual = eval("* {{ active=false }} OR * {{ term=\"clin find\" }}");
+		Expression expected = Expressions.builder()
+			.should(SnomedDocument.Expressions.inactive())
+			.should(SnomedDocument.Expressions.ids(Collections.emptySet()))
+			.build();
+		assertEquals(expected, actual);
+	}
+	
+	@Test
+	public void filterMultiDomainQueryExclusion() throws Exception {
+		Expression actual = eval("* {{ active=false }} MINUS * {{ term=\"clin find\" }}");
+		Expression expected = Expressions.builder()
+			.filter(SnomedDocument.Expressions.inactive())
+			.mustNot(SnomedDocument.Expressions.ids(Collections.emptySet()))
+			.build();
+		assertEquals(expected, actual);
+	}
+	
+	@Test
+	public void filterPreferredIn() throws Exception {
+		indexRevision(MAIN, SnomedDescriptionIndexEntry.builder()
+			.id(generateDescriptionId())
+			.active(true)
+			.moduleId(Concepts.MODULE_SCT_CORE)
+			.term("Clinical finding")
+			.conceptId(Concepts.ROOT_CONCEPT)
+			.typeId(Concepts.TEXT_DEFINITION)
+			.preferredIn(ImmutableSet.of(Concepts.REFSET_LANGUAGE_TYPE_UK))
+			.acceptableIn(ImmutableSet.of(Concepts.REFSET_LANGUAGE_TYPE_US))
+			.build());
+		
+		indexRevision(MAIN, SnomedDescriptionIndexEntry.builder()
+			.id(generateDescriptionId())
+			.active(true)
+			.moduleId(Concepts.MODULE_SCT_CORE)
+			.term("Clinical finding")
+			.conceptId(Concepts.SUBSTANCE)
+			.typeId(Concepts.TEXT_DEFINITION)
+			.preferredIn(ImmutableSet.of(Concepts.REFSET_LANGUAGE_TYPE_US))
+			.acceptableIn(ImmutableSet.of(Concepts.REFSET_LANGUAGE_TYPE_UK))
+			.build());
+		
+		final Expression actual = eval("* {{ preferredIn = " + Concepts.REFSET_LANGUAGE_TYPE_UK + " }}");
+		final Expression expected = SnomedDocument.Expressions.ids(ImmutableList.of(Concepts.ROOT_CONCEPT));
+		assertEquals(expected, actual);
+	}
+	
+	@Test
+	public void filterAcceptableIn() throws Exception {
+		indexRevision(MAIN, SnomedDescriptionIndexEntry.builder()
+			.id(generateDescriptionId())
+			.active(true)
+			.moduleId(Concepts.MODULE_SCT_CORE)
+			.term("Clinical finding")
+			.conceptId(Concepts.ROOT_CONCEPT)
+			.typeId(Concepts.TEXT_DEFINITION)
+			.preferredIn(ImmutableSet.of(Concepts.REFSET_LANGUAGE_TYPE_UK))
+			.acceptableIn(ImmutableSet.of(Concepts.REFSET_LANGUAGE_TYPE_US))
+			.build());
+		
+		indexRevision(MAIN, SnomedDescriptionIndexEntry.builder()
+			.id(generateDescriptionId())
+			.active(true)
+			.moduleId(Concepts.MODULE_SCT_CORE)
+			.term("Clinical finding")
+			.conceptId(Concepts.SUBSTANCE)
+			.typeId(Concepts.TEXT_DEFINITION)
+			.preferredIn(ImmutableSet.of(Concepts.REFSET_LANGUAGE_TYPE_US))
+			.acceptableIn(ImmutableSet.of(Concepts.REFSET_LANGUAGE_TYPE_UK))
+			.build());
+		
+		final Expression actual = eval("* {{ acceptableIn = " + Concepts.REFSET_LANGUAGE_TYPE_UK + " }}");
+		final Expression expected = SnomedDocument.Expressions.ids(ImmutableList.of(Concepts.SUBSTANCE));
+		assertEquals(expected, actual);
+	}
+	
+	@Test
+	public void filterLanguageRefSet() throws Exception {
+		indexRevision(MAIN, SnomedDescriptionIndexEntry.builder()
+				.id(generateDescriptionId())
+				.active(true)
+				.moduleId(Concepts.MODULE_SCT_CORE)
+				.term("Clinical finding")
+				.conceptId(Concepts.ROOT_CONCEPT)
+				.typeId(Concepts.TEXT_DEFINITION)
+				.preferredIn(ImmutableSet.of(Concepts.REFSET_LANGUAGE_TYPE_UK))
+				.acceptableIn(ImmutableSet.of(Concepts.REFSET_LANGUAGE_TYPE_US))
+				.build());
+		
+		indexRevision(MAIN, SnomedDescriptionIndexEntry.builder()
+				.id(generateDescriptionId())
+				.active(true)
+				.moduleId(Concepts.MODULE_SCT_CORE)
+				.term("Clinical finding")
+				.conceptId(Concepts.SUBSTANCE)
+				.typeId(Concepts.TEXT_DEFINITION)
+				.preferredIn(ImmutableSet.of(Concepts.REFSET_LANGUAGE_TYPE_US))
+				.acceptableIn(ImmutableSet.of(Concepts.REFSET_LANGUAGE_TYPE_UK))
+				.build());
+		
+		final Expression actual = eval("* {{ languageRefSetId = " + Concepts.REFSET_LANGUAGE_TYPE_UK + " }}");
+		final Expression expected = SnomedDocument.Expressions.ids(ImmutableSet.of(Concepts.ROOT_CONCEPT, Concepts.SUBSTANCE));
+		assertEquals(expected, actual);
+	}
+	
+	@Test(expected = BadRequestException.class)
+	public void filterInvalidLanguageCode() throws Exception {
+		eval("* {{ language = \"en-sg\" }}");
+	}
+	
+	@Test
+	public void filterLanguageCode() throws Exception {
+		indexRevision(MAIN, SnomedDescriptionIndexEntry.builder()
+				.id(generateDescriptionId())
+				.active(true)
+				.moduleId(Concepts.MODULE_SCT_CORE)
+				.term("Clinical finding")
+				.conceptId(Concepts.ROOT_CONCEPT)
+				.typeId(Concepts.TEXT_DEFINITION)
+				.languageCode("en")
+				.build());
+		
+		Expression actual = eval("* {{ language = en }}");
+		Expression expected = SnomedDocument.Expressions.ids(ImmutableSet.of(Concepts.ROOT_CONCEPT));
+		assertEquals(expected, actual);
+	}
+	
+	@Test
+	public void filterCaseSignificanceId() throws Exception {
+		indexRevision(MAIN, SnomedDescriptionIndexEntry.builder()
+				.id(generateDescriptionId())
+				.active(true)
+				.moduleId(Concepts.MODULE_SCT_CORE)
+				.term("Clinical finding")
+				.conceptId(Concepts.ROOT_CONCEPT)
+				.typeId(Concepts.TEXT_DEFINITION)
+				.languageCode("en")
+				.caseSignificanceId(Concepts.ENTIRE_TERM_CASE_INSENSITIVE)
+				.build());
+		
+		Expression actual = eval("* {{ caseSignificanceId = " + Concepts.ENTIRE_TERM_CASE_INSENSITIVE + " }}");
+		Expression expected = SnomedDocument.Expressions.ids(ImmutableSet.of(Concepts.ROOT_CONCEPT));
+		assertEquals(expected, actual);
+	}
+
 	/**
 	 * Generates the following test fixtures:
 	 * <ul>
